+++ class4js.js
@@ -627,11 +627,14 @@ var TypeBuilder = Object.create(Object.prototype, {
    */
   getPropertyDescriptor: {
     value: function (object, propertyName) {
-      var descriptor = Object.getOwnPropertyDescriptor(object, propertyName); 
-      if (!descriptor) {
-        return TypeBuilder.getPropertyDescriptor(Object.getPrototypeOf(object), propertyName);
+      if (object && propertyName) {
+        var descriptor = Object.getOwnPropertyDescriptor(object, propertyName); 
+        if (!descriptor) {
+          return TypeBuilder.getPropertyDescriptor(Object.getPrototypeOf(object), propertyName);
+        }
+        return descriptor;
       }
-      return descriptor;
+      return null;
     },
     writable: false,
     enumerable: true,
@@ -883,17 +886,20 @@ var Class = Object.create(Object.prototype, {
         if (parent && parent !== Object.prototype) {
           Class.initialize(instance, parent, args);
         }
-      }
-      if (prototype && prototype.hasOwnProperty('__construct__')) {
-        if (typeof prototype['__construct__'] === 'function') {
-          if (args && args.length == 1 && TypeBuilder.isObjectInitializer(args[0])) {
-            prototype['__construct__'].apply(instance);
-            ObjectFactory.initialize(instance, args[0]);
+        if (prototype.hasOwnProperty('__construct__')) {
+          if (typeof prototype['__construct__'] === 'function') {
+            if (args && args.length == 1 && TypeBuilder.isObjectInitializer(args[0])) {
+              prototype['__construct__'].apply(instance); 
+            } else {
+              prototype['__construct__'].apply(instance, args);
+            }
           } else {
-            prototype['__construct__'].apply(instance, args);
+            throw new TypeException("Class member's '__construct__' type is invalid");
           }
-        } else {
-          throw new TypeException("Class member's '__construct__' type is invalid");
+        }
+        if (Object.getPrototypeOf(instance) == prototype && args 
+            && args.length == 1 && TypeBuilder.isObjectInitializer(args[0])) {
+          ObjectFactory.initialize(instance, args[0]);
         }
       }
     },
@@ -1912,10 +1918,11 @@ var Proxy = Object.create(Object.prototype, {
    * @method create
    * @param {Object} type
    * @param {IInterceptor[]} interceptors
+   * @param {Array} args
    * @returns {Object}
    */
   create: {
-    value: function (type, interceptors) {
+    value: function (type, interceptors, args) {
       if (type) {
         if (interceptors) {
           if (typeof interceptors === 'function' || Interface.instanceOf(interceptors, IInterceptor)) {
@@ -1932,10 +1939,13 @@ var Proxy = Object.create(Object.prototype, {
         } else {
           throw new TypeException("Interceptor is not set"); 
         }
+        if (args && !(args instanceof Array)) {
+          args = [ args ];
+        }
         if (typeof type === 'function') {
-          return Proxy.__createClassProxy(type, interceptors);  
+          return Proxy.__createClassProxy(type, interceptors, args);  
         } else {
-          return Proxy.__createInterfaceProxy(type, interceptors); 
+          return Proxy.__createInterfaceProxy(type, interceptors, args); 
         }
       } else {
         throw new TypeException("Type is not set");
@@ -1972,21 +1982,27 @@ var Proxy = Object.create(Object.prototype, {
    * @returns {Object}
    */
   __createInterfaceProxy: {
-    value: function (type, interceptors) {
-      var proxy = Object.create(Object.prototype);
-      var descriptor;
+    value: function (type, interceptors, args) {
+      var constructor = function () {
+        Class.includeExtensions(this);
+        Class.initialize(this, Object.getPrototypeOf(this), args);
+        Object.seal(this);
+      };
+      constructor.prototype = Object.create(Object.prototype);
+      Proxy.__interceptClassConstructor(type, constructor.prototype, interceptors);
       for (var propertyName in type) {
         if (TypeBuilder.isPublic(propertyName)) {
-          descriptor = TypeBuilder.getPropertyDescriptor(type, propertyName);
+          var descriptor = TypeBuilder.getPropertyDescriptor(type, propertyName);
           if (TypeBuilder.isProperty(descriptor)) {
-            Proxy.__intercepProperty(proxy, propertyName, descriptor, interceptors, interceptors);
+            Proxy.__intercepProperty(constructor.prototype, propertyName, descriptor, interceptors, interceptors);
           } else if (TypeBuilder.isMethod(descriptor['value'])) {
-            Proxy.__interceptMethod(proxy, propertyName, descriptor, interceptors);
+            Proxy.__interceptMethod(constructor.prototype, propertyName, descriptor, interceptors);
           }
         }
       }
-      Object.seal(proxy);
-      return proxy;
+      Object.seal(constructor);
+      Object.seal(constructor.prototype);
+      return new constructor();
     },
     writable: false,
     enumerable: false,
@@ -2000,12 +2016,13 @@ var Proxy = Object.create(Object.prototype, {
    * @method __createClassProxy
    * @param {Object} type
    * @param {IInterceptor[]} interceptors
+   * @param {Array} args
    */
   __createClassProxy: {
-    value: function (type, interceptors) {
+    value: function (type, interceptors, args) {
       var constructor = function () {
         Class.includeExtensions(this);
-        Class.initialize(this, Object.getPrototypeOf(this), arguments);
+        Class.initialize(this, Object.getPrototypeOf(this), args);
         Object.seal(this);
       };
       constructor.prototype = Object.create(type.prototype);
@@ -2014,6 +2031,8 @@ var Proxy = Object.create(Object.prototype, {
         Proxy.__intercepClassMember(type.prototype, constructor.prototype, 
           propertyName, interceptors);
       }
+      Object.seal(constructor);
+      Object.seal(constructor.prototype);
       return new constructor();
     },
     writable: false,
@@ -2033,9 +2052,7 @@ var Proxy = Object.create(Object.prototype, {
   __interceptClassConstructor: {
     value: function (source, target, interceptors) {
       var descriptor = TypeBuilder.getPropertyDescriptor(source, '__construct__');
-      if (descriptor) {
-        Proxy.__interceptConstructor(target, descriptor, interceptors);
-      }
+      Proxy.__interceptConstructor(target, descriptor, interceptors);
     },
     writable: false,
     enumerable: true,
@@ -2103,9 +2120,11 @@ var Proxy = Object.create(Object.prototype, {
           arguments, interceptors);
         return invocation.procceed();
       };
-      TypeBuilder.addMethod(constructor, 'toString', function () {
-        return descriptor['value'].toString();
-      });
+      if (descriptor) {
+        TypeBuilder.addMethod(constructor, 'toString', function () {
+          return descriptor['value'].toString();
+        });
+      }
       TypeBuilder.addConstructor(proxy, '__construct__', constructor);
     },
     writable: false,